Renames package of viewmodel to domain
